feat: watch together

+++ source/app/src/main/java/com/movie_hub/android/data/model/mqtt/RoomStateModel.java
@@ -7,6 +7,6 @@ public class RoomStateModel {
     String subCmd;
 
     boolean isPlay;
-    long currentPositionMovie;
+    double currentPositionMovie;
     double playSpeed;
 }
